+++ preset/interceptors.js
@@ -259,5 +259,34 @@ function commonParallelRequest(option = {}) {
     };
 }
 
-export { commonCancelRequest, commonParallelRequest, commonThrottleRequest };
+/**超时 */
+function commonTimeoutRequest(option = {}) {
+    const { ms = 30000 } = option;
+    return {
+        install(interceptors) {
+            interceptors.request.use((url, init) => {
+                const abortController = new AbortController();
+                init.signal = abortController.signal;
+                const _commonTimeoutRequest = {
+                    controller: abortController,
+                    timeout: window.setTimeout(() => {
+                        abortController.abort();
+                        console.warn(`commonTimeoutRequest: timeout(${ms}). url:${url}, init:`, init);
+                    }, ms),
+                };
+                init._commonTimeoutRequest = _commonTimeoutRequest;
+                return init;
+            });
+            interceptors.response.use((data, { init }) => {
+                window.clearTimeout(init._commonTimeoutRequest.timeout);
+                return data;
+            }, err => {
+                window.clearTimeout(err.init._commonTimeoutRequest.timeout);
+                return Promise.reject(err);
+            });
+        },
+    };
+}
+
+export { commonCancelRequest, commonParallelRequest, commonThrottleRequest, commonTimeoutRequest };
 //# sourceMappingURL=interceptors.js.map