+++ src/preset/interceptors/index.ts
@@ -1,3 +1,4 @@
 export { default as commonCancelRequest } from './commonCancelRequest.js';
 export { default as commonThrottleRequest } from './commonThrottleRequest.js';
 export { default as commonParallelRequest } from './commonParallelRequest';
+export { default as commonTimeoutRequest } from './commonTimeoutRequest';