+++ src/utils.ts
@@ -29,7 +29,6 @@ export function checkInterceptorsReturn(rejectedFuncReturn: any, type: 'request'
     return rejectedFuncReturn;
   } else {
     console.warn(`${type}.interceptor.use(onFulfilled, onRejected): onRejected not return Promise.`);
-    // reject()
     return Promise.reject(errObj);
   }
 }